// src/shared/types/generated/candidatures.ts
export type TriCandidature = "poste" | "entreprise" | "statut" | "date";